создание нового модуля и вынесение независой функции в него.
